Change: Move SortOrder to pontos.github.model.base
SortOrder is either desc or asc and can be used at several places in the
GitHub API. Therefore it is a base model class.
